Disable sonarscan for dependabot
Dependabot PRs don't have access to build secrets, such as the token we
need to use to authenticate with sonarcloud.

The current dependabot workflow requires us to manually check each PR to
ensure only the sonarcloud step failed (and that stuff like tests are
fine) which causes friction and toil.

This change will allow us to even go as far as automerging dependabot,
should we want to- our dependabot PRs will actually become useful and
friction-free